app_front/pyproject.toml
[project]
name = "app-front"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
package = false

[dependency-groups]
app-front = [
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "python-dotenv>=1.0.1",
]
